Extended use of public error types
- updated rows package to use the new public error types
- refactored arrow/column based rows to put errors into a separate file
