+++ internal/rows/arrowbased/arrowRows.go
@@ -2,37 +2,22 @@ package arrowbased
 
 import (
 	"bytes"
+	"context"
 	"database/sql/driver"
-	"encoding/json"
-	"strings"
 
-	"fmt"
 	"time"
 
 	"github.com/apache/arrow/go/v11/arrow"
 	"github.com/apache/arrow/go/v11/arrow/array"
 	"github.com/apache/arrow/go/v11/arrow/ipc"
+	dbsqlerr "github.com/databricks/databricks-sql-go/errors"
 	"github.com/databricks/databricks-sql-go/internal/cli_service"
 	"github.com/databricks/databricks-sql-go/internal/config"
-	dbsqlerr "github.com/databricks/databricks-sql-go/internal/errors"
+	dbsqlerr_int "github.com/databricks/databricks-sql-go/internal/errors"
 	"github.com/databricks/databricks-sql-go/internal/rows/rowscanner"
 	dbsqllog "github.com/databricks/databricks-sql-go/logger"
 	"github.com/pkg/errors"
 )
-
-var errArrowRowsUnsupportedNativeType = "databricks: arrow native values not yet supported for %s"
-var errArrowRowsUnsupportedWithHiveSchema = "databricks: arrow native values for %s require arrow schema"
-var errArrowRowsInvalidBatchIndex = "databricks: invalid arrow batch index index = %d, n = %d"
-var errArrowRowsNoArrowBatches = "databricks: result set contains 0 arrow batches"
-var errArrowRowsUnableToReadBatch = "databricks: unable to read arrow batch"
-var errArrowRowsNilArrowSchema = "databricks: nil arrow.Schema"
-var errArrowRowsUnableToWriteArrowSchema = "databricks: unable to write arrow schema"
-var errArrowRowsInvalidRowIndex = "databricks: row index %d is not contained in any arrow batch"
-var errArrowRowsInvalidDecimalType = "databricks: decimal type with no scale/precision"
-var errArrowRowsUnableToCreateDecimalType = "databricks: unable to create decimal type scale: %d, precision: %d"
-var errArrowRowsUnknownDBType = "databricks: unknown data type when converting to arrow type"
-var errArrowRowsUnhandledArrowType = "databricks: arrow row scanner unhandled type %s"
-var errArrowRowsDateTimeParse = "databrics: arrow row scanner failed to parse date/time"
 
 type recordReader interface {
 	NewRecordFromBytes(arrowSchemaBytes []byte, sparkArrowBatch sparkArrowBatch) (arrow.Record, error)
@@ -96,13 +81,14 @@ type arrowRowScanner struct {
 	*dbsqllog.DBSQLLogger
 
 	location *time.Location
+	ctx      context.Context
 }
 
 // Make sure arrowRowScanner fulfills the RowScanner interface
 var _ rowscanner.RowScanner = (*arrowRowScanner)(nil)
 
 // NewArrowRowScanner returns an instance of RowScanner which handles arrow format results
-func NewArrowRowScanner(resultSetMetadata *cli_service.TGetResultSetMetadataResp, rowSet *cli_service.TRowSet, cfg *config.Config, logger *dbsqllog.DBSQLLogger) (rowscanner.RowScanner, error) {
+func NewArrowRowScanner(resultSetMetadata *cli_service.TGetResultSetMetadataResp, rowSet *cli_service.TRowSet, cfg *config.Config, logger *dbsqllog.DBSQLLogger, ctx context.Context) (rowscanner.RowScanner, dbsqlerr.DatabricksError) {
 
 	// we take a passed in logger, rather than just using the global from dbsqllog, so that the containing rows
 	// instance can pass in a logger with context such as correlation ID and operation ID
@@ -124,21 +110,21 @@ func NewArrowRowScanner(resultSetMetadata *cli_service.TGetResultSetMetadataResp
 		// convert the TTableSchema to an arrow Schema
 		arrowSchema, err = tTableSchemaToArrowSchema(resultSetMetadata.Schema, &arrowConfig)
 		if err != nil {
-			logger.Err(err).Msg("databricks: arrow row scanner failed to convert schema")
-			return nil, err
+			logger.Err(err).Msg(errArrowRowsConvertSchema)
+			return nil, dbsqlerr_int.NewSystemFault(ctx, errArrowRowsConvertSchema, err)
 		}
 
 		// serialize the arrow schema
-		schemaBytes, err = getArrowSchemaBytes(arrowSchema)
+		schemaBytes, err = getArrowSchemaBytes(arrowSchema, ctx)
 		if err != nil {
-			logger.Err(err).Msg("databricks: arrow row scanner failed to serialize schema")
-			return nil, err
+			logger.Err(err).Msg(errArrowRowsSerializeSchema)
+			return nil, dbsqlerr_int.NewSystemFault(ctx, errArrowRowsSerializeSchema, err)
 		}
 	} else {
 		br := &chunkedByteReader{chunks: [][]byte{schemaBytes}}
 		rdr, err := ipc.NewReader(br)
 		if err != nil {
-			return nil, dbsqlerr.WrapErr(err, errArrowRowsUnableToReadBatch)
+			return nil, dbsqlerr_int.NewSystemFault(ctx, errArrowRowsUnableToReadBatch, err)
 		}
 		defer rdr.Release()
 
@@ -157,9 +143,8 @@ func NewArrowRowScanner(resultSetMetadata *cli_service.TGetResultSetMetadataResp
 	// time values from the server are returned as UTC with microsecond precision
 	ttsf, err := arrow.FixedWidthTypes.Timestamp_us.(*arrow.TimestampType).GetToTimeFunc()
 	if err != nil {
-		logger.Err(err).Msg("databricks: arrow row scanner failed getting toTimestamp function")
-		err = dbsqlerr.WrapErr(err, "databricks: arrow row scanner failed getting toTimestamp function")
-		return nil, err
+		logger.Err(err).Msg(errArrowRowsToTimestampFn)
+		return nil, dbsqlerr_int.NewSystemFault(ctx, errArrowRowsToTimestampFn, err)
 	}
 
 	arrowBatches := make([]sparkArrowBatch, len(rowSet.ArrowBatches))
@@ -195,6 +180,7 @@ func NewArrowRowScanner(resultSetMetadata *cli_service.TGetResultSetMetadataResp
 		colInfo:             colInfos,
 		DBSQLLogger:         logger,
 		location:            location,
+		ctx:                 ctx,
 	}
 
 	return rs, nil
@@ -236,7 +222,7 @@ var intervalTypes map[cli_service.TTypeId]struct{} = map[cli_service.TTypeId]str
 // a buffer held in dest.
 func (ars *arrowRowScanner) ScanRow(
 	destination []driver.Value,
-	rowIndex int64) error {
+	rowIndex int64) dbsqlerr.DatabricksError {
 
 	// load the error batch for the specified row, if necessary
 	err := ars.loadBatchFor(rowIndex)
@@ -268,12 +254,16 @@ func (ars *arrowRowScanner) ScanRow(
 			if (dbType == cli_service.TTypeId_DECIMAL_TYPE && ars.UseArrowNativeDecimal) ||
 				(isIntervalType(dbType) && ars.UseArrowNativeIntervalTypes) {
 				//	not yet fully supported
-				ars.Error().Msgf(errArrowRowsUnsupportedNativeType, dbType)
-				return errors.Errorf(errArrowRowsUnsupportedNativeType, dbType)
+				ars.Error().Msgf(errArrowRowsUnsupportedNativeType(dbType.String()))
+				return dbsqlerr_int.NewSystemFault(ars.ctx, errArrowRowsUnsupportedNativeType(dbType.String()), nil)
 			}
 
 			// get the value from the column values holder
-			destination[i], err = ars.columnValues[i].Value(rowInBatchIndex)
+			var err1 error
+			destination[i], err1 = ars.columnValues[i].Value(rowInBatchIndex)
+			if err1 != nil {
+				err = dbsqlerr_int.NewSystemFault(ars.ctx, errArrowRowsColumnValue(col.name), err1)
+			}
 		}
 	}
 
@@ -301,7 +291,7 @@ func countRows(rowSet *cli_service.TRowSet) int64 {
 }
 
 // loadBatchFor loads the batch containing the specified row if necessary
-func (ars *arrowRowScanner) loadBatchFor(rowIndex int64) error {
+func (ars *arrowRowScanner) loadBatchFor(rowIndex int64) dbsqlerr.DatabricksError {
 
 	// if we haven't loaded the initial batch or the row is not in the current batch
 	// we need to load a different batch
@@ -323,12 +313,13 @@ func (ars *arrowRowScanner) loadBatchFor(rowIndex int64) error {
 }
 
 // loadBatch loads the arrow batch at the specified index
-func (ars *arrowRowScanner) loadBatch(batchIndex int) error {
+func (ars *arrowRowScanner) loadBatch(batchIndex int) dbsqlerr.DatabricksError {
 	if ars == nil || ars.arrowBatches == nil {
 		if ars != nil {
 			ars.Error().Msg(errArrowRowsNoArrowBatches)
+			return dbsqlerr_int.NewSystemFault(ars.ctx, errArrowRowsNoArrowBatches, nil)
 		}
-		return errors.New(errArrowRowsNoArrowBatches)
+		return dbsqlerr_int.NewSystemFault(context.Background(), errArrowRowsNoArrowBatches, nil)
 	}
 
 	// if the batch already loaded we can just return
@@ -337,22 +328,22 @@ func (ars *arrowRowScanner) loadBatch(batchIndex int) error {
 	}
 
 	if batchIndex < 0 || batchIndex >= len(ars.arrowBatches) {
-		ars.Error().Msgf(errArrowRowsInvalidBatchIndex, batchIndex, len(ars.arrowBatches))
-		return errors.Errorf(errArrowRowsInvalidBatchIndex, batchIndex, len(ars.arrowBatches))
+		ars.Error().Msg(errArrowRowsInvalidBatchIndex(batchIndex, len(ars.arrowBatches)))
+		return dbsqlerr_int.NewSystemFault(ars.ctx, errArrowRowsInvalidBatchIndex(batchIndex, len(ars.arrowBatches)), nil)
 	}
 
 	// set up the column values containers
 	if ars.columnValues == nil {
 		err := ars.makeColumnValuesContainers(ars)
 		if err != nil {
-			return err
+			return dbsqlerr_int.NewSystemFault(ars.ctx, errArrowRowsMakeColumnValueContainers, err)
 		}
 	}
 
 	r, err := ars.NewRecordFromBytes(ars.arrowSchemaBytes, ars.arrowBatches[batchIndex])
 	if err != nil {
 		ars.Err(err).Msg(errArrowRowsUnableToReadBatch)
-		return err
+		return dbsqlerr_int.NewSystemFault(ars.ctx, errArrowRowsUnableToReadBatch, err)
 	}
 
 	defer r.Release()
@@ -383,16 +374,16 @@ func (ars *arrowRowScanner) loadBatch(batchIndex int) error {
 }
 
 // getArrowSchemaBytes returns the serialized schema in ipc format
-func getArrowSchemaBytes(schema *arrow.Schema) ([]byte, error) {
+func getArrowSchemaBytes(schema *arrow.Schema, ctx context.Context) ([]byte, dbsqlerr.DatabricksError) {
 	if schema == nil {
-		return nil, errors.New(errArrowRowsNilArrowSchema)
+		return nil, dbsqlerr_int.NewSystemFault(ctx, errArrowRowsNilArrowSchema, nil)
 	}
 
 	var output bytes.Buffer
 	w := ipc.NewWriter(&output, ipc.WithSchema(schema))
 	err := w.Close()
 	if err != nil {
-		return nil, dbsqlerr.WrapErr(err, errArrowRowsUnableToWriteArrowSchema)
+		return nil, dbsqlerr_int.NewSystemFault(ctx, errArrowRowsUnableToWriteArrowSchema, err)
 	}
 
 	arrowSchemaBytes := output.Bytes()
@@ -405,7 +396,7 @@ func getArrowSchemaBytes(schema *arrow.Schema) ([]byte, error) {
 }
 
 // rowIndexToBatchIndex returns the index of the batch containing the specified row
-func (ars *arrowRowScanner) rowIndexToBatchIndex(rowIndex int64) (int, error) {
+func (ars *arrowRowScanner) rowIndexToBatchIndex(rowIndex int64) (int, dbsqlerr.DatabricksError) {
 
 	for i := range ars.arrowBatches {
 		if ars.arrowBatches[i].contains(rowIndex) {
@@ -413,8 +404,8 @@ func (ars *arrowRowScanner) rowIndexToBatchIndex(rowIndex int64) (int, error) {
 		}
 	}
 
-	ars.Error().Msgf(errArrowRowsInvalidRowIndex, rowIndex)
-	return -1, errors.Errorf(errArrowRowsInvalidRowIndex, rowIndex)
+	ars.Error().Msg(errArrowRowsInvalidRowIndex(rowIndex))
+	return -1, dbsqlerr_int.NewSystemFault(ars.ctx, errArrowRowsInvalidRowIndex(rowIndex), nil)
 }
 
 // tTableSchemaToArrowSchema convers the TTableSchema retrieved by the thrift server into an arrow.Schema instance
@@ -486,7 +477,7 @@ func tColumnDescToArrowDataType(tColumnDesc *cli_service.TColumnDesc, arrowConfi
 
 			decimalType, err := arrow.NewDecimalType(arrow.DECIMAL128, precision, scale)
 			if err != nil {
-				return nil, dbsqlerr.WrapErr(err, fmt.Sprintf(errArrowRowsUnableToCreateDecimalType, scale, precision))
+				return nil, dbsqlerr_int.WrapErr(err, errArrowRowsUnableToCreateDecimalType(scale, precision))
 			}
 
 			return decimalType, nil
@@ -505,14 +496,14 @@ func tColumnDescToArrowDataType(tColumnDesc *cli_service.TColumnDesc, arrowConfi
 				return arrow.BinaryTypes.String, nil
 			}
 
-			return nil, errors.Errorf(errArrowRowsUnsupportedWithHiveSchema, rowscanner.GetDBTypeName(tColumnDesc))
+			return nil, errors.New(errArrowRowsUnsupportedWithHiveSchema(rowscanner.GetDBTypeName(tColumnDesc)))
 		} else if _, ok := intervalTypes[tType]; ok {
 			// if not using arrow native complex types thrift server returns strings
 			if !arrowConfig.UseArrowNativeIntervalTypes {
 				return arrow.BinaryTypes.String, nil
 			}
 
-			return nil, errors.Errorf(errArrowRowsUnsupportedWithHiveSchema, rowscanner.GetDBTypeName(tColumnDesc))
+			return nil, errors.New(errArrowRowsUnsupportedWithHiveSchema(rowscanner.GetDBTypeName(tColumnDesc)))
 		} else {
 			return nil, errors.New(errArrowRowsUnknownDBType)
 		}
@@ -562,110 +553,6 @@ func tColumnDescToArrowField(columnDesc *cli_service.TColumnDesc, arrowConfig *c
 	return arrowField, nil
 }
 
-// columnValues is the interface for accessing the values for a column
-type columnValues interface {
-	Value(int) (any, error)
-	IsNull(int) bool
-	Release()
-	SetValueArray(colData arrow.ArrayData) error
-}
-
-// a type constraint for the value types which we handle that are returned in the arrow records
-type valueTypes interface {
-	bool |
-		int8 |
-		int16 |
-		int32 |
-		int64 |
-		float32 |
-		float64 |
-		string |
-		[]byte
-}
-
-// a type constraint for the arrow array types which we handle that are returned in the arrow records
-type arrowArrayTypes interface {
-	*array.Boolean |
-		*array.Int8 |
-		*array.Int16 |
-		*array.Int32 |
-		*array.Int64 |
-		*array.Float32 |
-		*array.Float64 |
-		*array.String |
-		*array.Binary
-}
-
-// type constraint for wrapping arrow arrays
-type columnValuesHolder[T valueTypes] interface {
-	arrowArrayTypes
-	Value(int) T
-	IsNull(int) bool
-	Release()
-}
-
-// a generic container for the arrow arrays/value types we handle
-type columnValuesTyped[ValueHolderType columnValuesHolder[ValueType], ValueType valueTypes] struct {
-	holder ValueHolderType
-	foo    ValueType
-}
-
-// return the value for the specified row
-func (cv *columnValuesTyped[X, T]) Value(rowNum int) (any, error) {
-	return cv.holder.Value(rowNum), nil
-}
-
-// return true if the value at rowNum is null
-func (cv *columnValuesTyped[X, T]) IsNull(rowNum int) bool {
-	return cv.holder.IsNull(rowNum)
-}
-
-// release the the contained arrow array
-func (cv *columnValuesTyped[X, T]) Release() {
-	if cv.holder != nil {
-		cv.holder.Release()
-	}
-}
-
-func (cv *columnValuesTyped[X, T]) SetValueArray(colData arrow.ArrayData) error {
-	var colValsHolder columnValues = cv
-	switch t := any(cv.foo).(type) {
-	case bool:
-		colValsHolder.(*columnValuesTyped[*array.Boolean, bool]).holder = array.NewBooleanData(colData)
-
-	case int8:
-		colValsHolder.(*columnValuesTyped[*array.Int8, int8]).holder = array.NewInt8Data(colData)
-
-	case int16:
-		colValsHolder.(*columnValuesTyped[*array.Int16, int16]).holder = array.NewInt16Data(colData)
-
-	case int32:
-		colValsHolder.(*columnValuesTyped[*array.Int32, int32]).holder = array.NewInt32Data(colData)
-
-	case int64:
-		colValsHolder.(*columnValuesTyped[*array.Int64, int64]).holder = array.NewInt64Data(colData)
-
-	case float32:
-		colValsHolder.(*columnValuesTyped[*array.Float32, float32]).holder = array.NewFloat32Data(colData)
-
-	case float64:
-		colValsHolder.(*columnValuesTyped[*array.Float64, float64]).holder = array.NewFloat64Data(colData)
-
-	case string:
-		colValsHolder.(*columnValuesTyped[*array.String, string]).holder = array.NewStringData(colData)
-
-	case []byte:
-		colValsHolder.(*columnValuesTyped[*array.Binary, []byte]).holder = array.NewBinaryData(colData)
-
-	default:
-		return errors.Errorf(errArrowRowsUnhandledArrowType, t)
-	}
-
-	return nil
-}
-
-var _ columnValues = (*columnValuesTyped[*array.Int16, int16])(nil)
-
 type sparkRecordReader struct{}
 
 // Make sure sparkRecordReader fulfills the recordReader interface
@@ -680,13 +567,13 @@ func (srr sparkRecordReader) NewRecordFromBytes(arrowSchemaBytes []byte, sparkAr
 	br := &chunkedByteReader{chunks: [][]byte{arrowSchemaBytes, sparkArrowBatch.arrowRecordBytes}}
 	rdr, err := ipc.NewReader(br)
 	if err != nil {
-		return nil, dbsqlerr.WrapErr(err, errArrowRowsUnableToReadBatch)
+		return nil, err
 	}
 	defer rdr.Release()
 
 	r, err := rdr.Read()
 	if err != nil {
-		return nil, dbsqlerr.WrapErr(err, errArrowRowsUnableToReadBatch)
+		return nil, err
 	}
 
 	r.Retain()
@@ -815,347 +702,6 @@ func (vcm *arrowValueContainerMaker) makeColumnValueContainer(t arrow.DataType, 
 		return svc, nil
 
 	default:
-		return nil, errors.Errorf(errArrowRowsUnhandledArrowType, t.String())
+		return nil, errors.Errorf(errArrowRowsUnhandledArrowType(t.String()))
 	}
-}
-
-type listValueContainer struct {
-	listArray     array.ListLike
-	values        columnValues
-	complexValue  bool
-	listArrayType *arrow.ListType
-}
-
-func (lvc *listValueContainer) Value(i int) (any, error) {
-	if i < lvc.listArray.Len() {
-		r := "["
-		s, e := lvc.listArray.ValueOffsets(i)
-		len := int(e - s)
-
-		for i := 0; i < len; i++ {
-			val, err := lvc.values.Value(i + int(s))
-			if err != nil {
-				return nil, err
-			}
-
-			if !lvc.complexValue {
-				vb, err := marshal(val)
-				if err != nil {
-					return nil, err
-				}
-				r = r + string(vb)
-			} else {
-				r = r + val.(string)
-			}
-
-			if i < len-1 {
-				r = r + ","
-			}
-		}
-
-		r = r + "]"
-		return r, nil
-	}
-	return nil, nil
-}
-
-func (lvc *listValueContainer) IsNull(i int) bool {
-	return lvc.listArray.IsNull(i)
-}
-
-func (lvc *listValueContainer) Release() {
-	if lvc.listArray != nil {
-		lvc.listArray.Release()
-	}
-
-	if lvc.values != nil {
-		lvc.values.Release()
-	}
-}
-
-func (lvc *listValueContainer) SetValueArray(colData arrow.ArrayData) error {
-	lvc.listArray = array.NewListData(colData)
-	lvs := lvc.listArray.ListValues()
-	err := lvc.values.SetValueArray(lvs.Data())
-
-	return err
-}
-
-type mapValueContainer struct {
-	mapArray     *array.Map
-	keys         columnValues
-	values       columnValues
-	complexValue bool
-	mapArrayType *arrow.MapType
-}
-
-func (mvc *mapValueContainer) Value(i int) (any, error) {
-	if i < mvc.mapArray.Len() {
-		s, e := mvc.mapArray.ValueOffsets(i)
-		len := e - s
-		r := "{"
-		for i := int64(0); i < len; i++ {
-			k, err := mvc.keys.Value(int(i))
-			if err != nil {
-				return nil, err
-			}
-
-			v, err := mvc.values.Value(int(i))
-			if err != nil {
-				return nil, err
-			}
-
-			key, err := marshal(k)
-			if err != nil {
-				return nil, err
-			}
-
-			var b string
-			if mvc.complexValue {
-				b = v.(string)
-			} else {
-				vb, err := marshal(v)
-				if err != nil {
-					return nil, err
-				}
-				b = string(vb)
-			}
-
-			if !strings.HasPrefix(string(key), "\"") {
-				r = r + "\"" + string(key) + "\":"
-			} else {
-				r = r + string(key) + ":"
-			}
-
-			r = r + b
-			if i < len-1 {
-				r = r + ","
-			}
-		}
-		r = r + "}"
-
-		return r, nil
-	}
-	return nil, nil
-}
-
-func (mvc *mapValueContainer) IsNull(i int) bool {
-	return mvc.mapArray.IsNull(i)
-}
-
-func (mvc *mapValueContainer) Release() {
-	if mvc.mapArray != nil {
-		mvc.mapArray.Release()
-	}
-
-	if mvc.values != nil {
-		mvc.values.Release()
-	}
-
-	if mvc.keys != nil {
-		mvc.keys.Release()
-	}
-}
-
-func (mvc *mapValueContainer) SetValueArray(colData arrow.ArrayData) error {
-	mvc.mapArray = array.NewMapData(colData)
-	err := mvc.values.SetValueArray(mvc.mapArray.Items().Data())
-	if err != nil {
-		return err
-	}
-	err = mvc.keys.SetValueArray(mvc.mapArray.Keys().Data())
-
-	return err
-}
-
-type structValueContainer struct {
-	structArray     *array.Struct
-	fieldNames      []string
-	complexValue    []bool
-	fieldValues     []columnValues
-	structArrayType *arrow.StructType
-}
-
-func (svc *structValueContainer) Value(i int) (any, error) {
-	if i < svc.structArray.Len() {
-		r := "{"
-		for j := range svc.fieldValues {
-			r = r + "\"" + svc.fieldNames[j] + "\":"
-
-			v, err := svc.fieldValues[j].Value(int(i))
-			if err != nil {
-				return nil, err
-			}
-
-			var b string
-			if svc.complexValue[j] {
-				b = v.(string)
-			} else {
-				vb, err := marshal(v)
-				if err != nil {
-					return nil, err
-				}
-				b = string(vb)
-			}
-
-			r = r + b
-			if j < len(svc.fieldValues)-1 {
-				r = r + ","
-			}
-		}
-		r = r + "}"
-
-		return r, nil
-	}
-	return nil, nil
-}
-
-func (svc *structValueContainer) IsNull(i int) bool {
-	return svc.structArray.IsNull(i)
-}
-
-func (svc *structValueContainer) Release() {
-	if svc.structArray != nil {
-		svc.structArray.Release()
-	}
-
-	for i := range svc.fieldValues {
-		if svc.fieldValues[i] != nil {
-			svc.fieldValues[i].Release()
-		}
-	}
-}
-
-func (svc *structValueContainer) SetValueArray(colData arrow.ArrayData) error {
-	svc.structArray = array.NewStructData(colData)
-	for i := range svc.fieldValues {
-		err := svc.fieldValues[i].SetValueArray(svc.structArray.Field(i).Data())
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-type dateValueContainer struct {
-	dateArray *array.Date32
-	location  *time.Location
-}
-
-func (dvc *dateValueContainer) Value(i int) (any, error) {
-	d32 := dvc.dateArray.Value(i)
-
-	val := d32.ToTime().In(dvc.location)
-	return val, nil
-}
-
-func (dvc *dateValueContainer) IsNull(i int) bool {
-	return dvc.dateArray.IsNull(i)
-}
-
-func (dvc *dateValueContainer) Release() {
-	if dvc.dateArray != nil {
-		dvc.dateArray.Release()
-	}
-}
-
-func (dvc *dateValueContainer) SetValueArray(colData arrow.ArrayData) error {
-	dvc.dateArray = array.NewDate32Data(colData)
-	return nil
-}
-
-type timestampValueContainer struct {
-	timeArray     *array.Timestamp
-	location      *time.Location
-	toTimestampFn func(arrow.Timestamp) time.Time
-}
-
-func (tvc *timestampValueContainer) Value(i int) (any, error) {
-	ats := tvc.timeArray.Value(i)
-	val := tvc.toTimestampFn(ats).In(tvc.location)
-
-	return val, nil
-}
-
-func (tvc *timestampValueContainer) IsNull(i int) bool {
-	return tvc.timeArray.IsNull(i)
-}
-
-func (tvc *timestampValueContainer) Release() {
-	if tvc.timeArray != nil {
-		tvc.timeArray.Release()
-	}
-}
-
-func (tvc *timestampValueContainer) SetValueArray(colData arrow.ArrayData) error {
-	tvc.timeArray = array.NewTimestampData(colData)
-	return nil
-}
-
-type timestampStringValueContainer struct {
-	timeStringArray *array.String
-	location        *time.Location
-	fieldName       string
-	*dbsqllog.DBSQLLogger
-}
-
-func (tvc *timestampStringValueContainer) Value(i int) (any, error) {
-	sv := tvc.timeStringArray.Value(i)
-	val, err := rowscanner.HandleDateTime(sv, "TIMESTAMP", tvc.fieldName, tvc.location)
-	if err != nil {
-		tvc.Err(err).Msg(errArrowRowsDateTimeParse)
-	}
-
-	return val, nil
-}
-
-func (tvc *timestampStringValueContainer) IsNull(i int) bool {
-	return tvc.timeStringArray.IsNull(i)
-}
-
-func (tvc *timestampStringValueContainer) Release() {
-	if tvc.timeStringArray != nil {
-		tvc.timeStringArray.Release()
-	}
-}
-
-func (tvc *timestampStringValueContainer) SetValueArray(colData arrow.ArrayData) error {
-	tvc.timeStringArray = array.NewStringData(colData)
-	return nil
-}
-
-type decimal128Container struct {
-	decimalArray *array.Decimal128
-	scale        int32
-}
-
-func (tvc *decimal128Container) Value(i int) (any, error) {
-	dv := tvc.decimalArray.Value(i)
-	fv := dv.ToFloat64(tvc.scale)
-	return fv, nil
-}
-
-func (tvc *decimal128Container) IsNull(i int) bool {
-	return tvc.decimalArray.IsNull(i)
-}
-
-func (tvc *decimal128Container) Release() {
-	if tvc.decimalArray != nil {
-		tvc.decimalArray.Release()
-	}
-}
-
-func (tvc *decimal128Container) SetValueArray(colData arrow.ArrayData) error {
-	tvc.decimalArray = array.NewDecimal128Data(colData)
-	return nil
-}
-
-func marshal(val any) ([]byte, error) {
-	if t, ok := val.(time.Time); ok {
-		s := "\"" + t.String() + "\""
-		return []byte(s), nil
-	}
-	vb, err := json.Marshal(val)
-	return vb, err
 }
